+++ src/main/java/wooteco/subway/line/application/LineService.java
@@ -11,6 +11,7 @@ import wooteco.subway.line.domain.Section;
 import wooteco.subway.line.dto.LineRequest;
 import wooteco.subway.line.dto.LineResponse;
 import wooteco.subway.line.dto.SectionRequest;
+import wooteco.subway.line.exception.DuplicatedLineException;
 import wooteco.subway.station.application.StationService;
 import wooteco.subway.station.domain.Station;
