+++ src/main/java/wooteco/subway/line/exception/DuplicatedLineException.java
@@ -1,4 +1,4 @@
-package wooteco.subway.line.application;
+package wooteco.subway.line.exception;
 
 import org.springframework.http.HttpStatus;
 import org.springframework.web.bind.annotation.ResponseStatus;
