+++ src/main/java/wooteco/subway/member/application/MemberService.java
@@ -7,6 +7,7 @@ import wooteco.subway.member.domain.LoginMember;
 import wooteco.subway.member.domain.Member;
 import wooteco.subway.member.dto.MemberRequest;
 import wooteco.subway.member.dto.MemberResponse;
+import wooteco.subway.member.exception.DuplicatedMemberException;
 
 @Service
 @Transactional(readOnly = true)
