+++ src/main/java/wooteco/subway/member/exception/DuplicatedMemberException.java
@@ -1,12 +1,12 @@
-package wooteco.subway.station.application;
+package wooteco.subway.member.exception;
 
 import org.springframework.http.HttpStatus;
 import org.springframework.web.bind.annotation.ResponseStatus;
 
 @ResponseStatus(HttpStatus.BAD_REQUEST)
-public class InvalidDeletionException extends RuntimeException {
+public class DuplicatedMemberException extends RuntimeException {
 
-    public InvalidDeletionException(String message) {
+    public DuplicatedMemberException(String message) {
         super(message);
     }
 }
