+++ src/main/java/wooteco/subway/path/exception/InvalidPathException.java
@@ -1,13 +1,12 @@
-package wooteco.subway.station.application;
+package wooteco.subway.path.exception;
 
 import org.springframework.http.HttpStatus;
 import org.springframework.web.bind.annotation.ResponseStatus;
 
 @ResponseStatus(HttpStatus.BAD_REQUEST)
-public class NotFoundException extends RuntimeException {
+public class InvalidPathException extends RuntimeException {
 
-    public NotFoundException(String message) {
+    public InvalidPathException(String message) {
         super(message);
     }
-
 }
