+++ src/main/java/wooteco/subway/station/application/StationService.java
@@ -9,6 +9,9 @@ import wooteco.subway.station.dao.StationDao;
 import wooteco.subway.station.domain.Station;
 import wooteco.subway.station.dto.StationRequest;
 import wooteco.subway.station.dto.StationResponse;
+import wooteco.subway.station.exception.DuplicatedStationException;
+import wooteco.subway.station.exception.InvalidDeletionException;
+import wooteco.subway.station.exception.NotFoundException;
 
 @Service
 @Transactional(readOnly = true)
