+++ src/main/java/wooteco/subway/station/exception/InvalidDeletionException.java
@@ -1,12 +1,12 @@
-package wooteco.subway.path.application;
+package wooteco.subway.station.exception;
 
 import org.springframework.http.HttpStatus;
 import org.springframework.web.bind.annotation.ResponseStatus;
 
 @ResponseStatus(HttpStatus.BAD_REQUEST)
-public class InvalidPathException extends RuntimeException {
+public class InvalidDeletionException extends RuntimeException {
 
-    public InvalidPathException(String message) {
+    public InvalidDeletionException(String message) {
         super(message);
     }
 }
