+++ src/main/java/wooteco/subway/station/exception/NotFoundException.java
@@ -1,12 +1,13 @@
-package wooteco.subway.member.application;
+package wooteco.subway.station.exception;
 
 import org.springframework.http.HttpStatus;
 import org.springframework.web.bind.annotation.ResponseStatus;
 
 @ResponseStatus(HttpStatus.BAD_REQUEST)
-public class DuplicatedMemberException extends RuntimeException {
+public class NotFoundException extends RuntimeException {
 
-    public DuplicatedMemberException(String message) {
+    public NotFoundException(String message) {
         super(message);
     }
+
 }
