+++ src/test/java/wooteco/subway/member/application/MemberServiceTest.java
@@ -17,6 +17,7 @@ import wooteco.subway.member.domain.LoginMember;
 import wooteco.subway.member.domain.Member;
 import wooteco.subway.member.dto.MemberRequest;
 import wooteco.subway.member.dto.MemberResponse;
+import wooteco.subway.member.exception.DuplicatedMemberException;
 
 @MockitoSettings
 class MemberServiceTest {
