+++ src/test/java/wooteco/subway/station/application/StationServiceTest.java
@@ -18,6 +18,9 @@ import wooteco.subway.station.dao.StationDao;
 import wooteco.subway.station.domain.Station;
 import wooteco.subway.station.dto.StationRequest;
 import wooteco.subway.station.dto.StationResponse;
+import wooteco.subway.station.exception.DuplicatedStationException;
+import wooteco.subway.station.exception.InvalidDeletionException;
+import wooteco.subway.station.exception.NotFoundException;
 
 @MockitoSettings
 class StationServiceTest {
